a billion modifications to make it work with the actual game, and splitting messages into ingoing and outgoing

+++ ArxLibertatisServer/ServerState.cs
@@ -2,6 +2,6 @@
 {
     public class ServerState
     {
-        public long level;
+        public int level;
     }
 }
